[S07P22A604-227] feat: 재료 가격 정보 크롤러 추가

+++ molaeng_back/src/main/java/idle/molaeng_back/recipe/model/response/IngredientResponse.java
@@ -49,7 +49,7 @@ public class IngredientResponse {
         @ApiParam("재료명")
         private String ingredientName;
         @ApiParam("기준 무게 당 재료 가격")
-        private int ingredientPrice;
+        private double ingredientPrice;
         @ApiParam("기준 무게")
         private int ingredientWeight;
         @ApiParam("기준 무게의 무게단위")
